Criando estrutura

+++ src/main/java/br/com/telematica/siloapi/repository/UsuarioRepository.java
@@ -1,13 +1,11 @@
 package br.com.telematica.siloapi.repository;
 
-import java.math.BigInteger;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import br.com.telematica.siloapi.model.enttity.UsuarioEntity;
+import br.com.telematica.siloapi.model.entity.UsuarioEntity;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<UsuarioEntity, BigInteger>{
-    UsuarioEntity findByUsulog(String login);
+public interface UsuarioRepository extends JpaRepository<UsuarioEntity, Integer> {
+	UsuarioEntity findByUsulog(String login);
 }
